much needed renaming
ClickableTrigger -> Selectable
CursorObject -> Pettable
PettingManager -> CursorManager
